fix: fix errors
