+++ docs/reference/configuration.md
@@ -689,7 +689,7 @@ A MongoDB sink Firehose \(`SINK_TYPE`= `mongodb` \) requires the following varia
 
 MongoDB connection URL/URLs to connect. Multiple URLs could be given in a comma separated format.
 
-* Example value: `localhost1:9200`
+* Example value: `localhost:27017`
 * Type: `required`
 
 ### `SINK_MONGO_DB_NAME`
