+++ src/test/java/io/odpf/firehose/sink/mongodb/util/MongoSinkFactoryUtilTest.java
@@ -1,96 +1,123 @@
 package io.odpf.firehose.sink.mongodb.util;
 
-import io.odpf.firehose.config.MongoSinkConfig;
+import com.mongodb.ServerAddress;
 import io.odpf.firehose.metrics.Instrumentation;
-import org.aeonbits.owner.ConfigFactory;
 import org.junit.Before;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.ExpectedException;
 import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
 
-import java.util.HashMap;
+import java.util.List;
 
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.MockitoAnnotations.initMocks;
+import static org.junit.Assert.assertEquals;
 
 public class MongoSinkFactoryUtilTest {
 
-    private MongoSinkConfig mongoSinkConfig;
-    private HashMap<String, String> config = new HashMap<>();
+    @Rule
+    public ExpectedException thrown = ExpectedException.none();
 
     @Mock
     private Instrumentation instrumentation;
 
-
     @Before
-    public void setup() {
-        initMocks(this);
-
-        config.put("SINK_MONGO_CONNECTION_URLS", "localhost:8080");
-        config.put("SINK_MONGO_DB_NAME", "sampleDatabase");
-        config.put("SINK_MONGO_PRIMARY_KEY", "customer_id");
-        config.put("SINK_MONGO_INPUT_MESSAGE_TYPE", "JSON");
-        config.put("SINK_MONGO_COLLECTION_NAME", "customers");
-        config.put("SINK_MONGO_REQUEST_TIMEOUT_MS", "8277");
-        config.put("SINK_MONGO_RETRY_STATUS_CODE_BLACKLIST", "11000");
-        config.put("SINK_MONGO_MODE_UPDATE_ONLY_ENABLE", "true");
-        config.put("SINK_MONGO_AUTH_ENABLE", "false");
-
+    public void setUp() {
+        MockitoAnnotations.initMocks(this);
     }
 
-
     @Test
-    public void shouldLogMongoSinkConfig() {
-        mongoSinkConfig = ConfigFactory.create(MongoSinkConfig.class, config);
+    public void shouldThrowIllegalArgumentExceptionForEmptyMongoConnectionURLs() {
+        thrown.expect(IllegalArgumentException.class);
+        thrown.expectMessage("SINK_MONGO_CONNECTION_URLS is empty or null");
 
-        MongoSinkFactoryUtil.logMongoConfig(mongoSinkConfig, instrumentation);
-        verify(instrumentation, times(1)).logDebug(any());
+        MongoSinkFactoryUtil.getServerAddresses("", instrumentation);
 
     }
 
     @Test
-    public void shouldLogMongoSinkConfigWithCorrectMessageWhenAuthDisabled() {
-        mongoSinkConfig = ConfigFactory.create(MongoSinkConfig.class, config);
-
-        MongoSinkFactoryUtil.logMongoConfig(mongoSinkConfig, instrumentation);
-        verify(instrumentation, times(1))
-                .logDebug("\n\tMONGO connection urls: localhost:8080"
-                        + "\n\tMONGO Database name: sampleDatabase"
-                        + "\n\tMONGO Primary Key: customer_id"
-                        + "\n\tMONGO input message type: JSON"
-                        + "\n\tMONGO Collection Name: customers"
-                        + "\n\tMONGO request timeout in ms: 8277"
-                        + "\n\tMONGO retry status code blacklist: 11000"
-                        + "\n\tMONGO update only mode: true"
-                        + "\n\tMONGO Authentication Enable: false"
-                        + "\n\tMONGO Authentication Username: null"
-                        + "\n\tMONGO Authentication Database: null");
-
+    public void shouldThrowIllegalArgumentExceptionWhenServerPortInvalid() {
+        String mongoConnectionURLs = "localhost:qfb";
+        thrown.expect(IllegalArgumentException.class);
+        MongoSinkFactoryUtil.getServerAddresses(mongoConnectionURLs, instrumentation);
     }
 
     @Test
-    public void shouldLogMongoSinkConfigWithCorrectMessageWhenAuthEnabled() {
+    public void shouldThrowIllegalArgumentExceptionForNullMongoConnectionURLs() {
+        thrown.expect(IllegalArgumentException.class);
+        thrown.expectMessage("SINK_MONGO_CONNECTION_URLS is empty or null");
 
-        config.put("SINK_MONGO_AUTH_ENABLE", "true");
-        config.put("SINK_MONGO_AUTH_USERNAME", "john");
-        config.put("SINK_MONGO_AUTH_DB", "agents");
+        MongoSinkFactoryUtil.getServerAddresses(null, instrumentation);
+    }
 
-        mongoSinkConfig = ConfigFactory.create(MongoSinkConfig.class, config);
+    @Test
+    public void shouldThrowIllegalArgumentExceptionForEmptyHost() {
+        String mongoConnectionURLs = ":1000";
+        thrown.expect(IllegalArgumentException.class);
+        MongoSinkFactoryUtil.getServerAddresses(mongoConnectionURLs, instrumentation);
+    }
 
-        MongoSinkFactoryUtil.logMongoConfig(mongoSinkConfig, instrumentation);
-        verify(instrumentation, times(1))
-                .logDebug("\n\tMONGO connection urls: localhost:8080"
-                        + "\n\tMONGO Database name: sampleDatabase"
-                        + "\n\tMONGO Primary Key: customer_id"
-                        + "\n\tMONGO input message type: JSON"
-                        + "\n\tMONGO Collection Name: customers"
-                        + "\n\tMONGO request timeout in ms: 8277"
-                        + "\n\tMONGO retry status code blacklist: 11000"
-                        + "\n\tMONGO update only mode: true"
-                        + "\n\tMONGO Authentication Enable: true"
-                        + "\n\tMONGO Authentication Username: john"
-                        + "\n\tMONGO Authentication Database: agents");
+    @Test
+    public void shouldThrowIllegalArgumentExceptionForEmptyPort() {
+        String mongoConnectionURLs = "localhost:";
+        thrown.expect(IllegalArgumentException.class);
 
+        MongoSinkFactoryUtil.getServerAddresses(mongoConnectionURLs, instrumentation);
+    }
+
+    @Test
+    public void shouldGetServerAddressesForValidMongoConnectionURLs() {
+        String mongoConnectionURLs = "localhost_1:1000,localhost_2:1000";
+        List<ServerAddress> serverAddresses = MongoSinkFactoryUtil.getServerAddresses(mongoConnectionURLs, instrumentation);
+
+        assertEquals("localhost_1", serverAddresses.get(0).getHost());
+        assertEquals(1000, serverAddresses.get(0).getPort());
+        assertEquals("localhost_2", serverAddresses.get(1).getHost());
+        assertEquals(1000, serverAddresses.get(1).getPort());
+    }
+
+    @Test
+    public void shouldGetServerAddressesForValidMongoConnectionURLsWithSpacesInBetween() {
+        String mongoConnectionURLs = " localhost_1: 1000,  localhost_2:1000";
+        List<ServerAddress> serverAddresses = MongoSinkFactoryUtil.getServerAddresses(mongoConnectionURLs, instrumentation);
+
+        assertEquals("localhost_1", serverAddresses.get(0).getHost());
+        assertEquals(1000, serverAddresses.get(0).getPort());
+        assertEquals("localhost_2", serverAddresses.get(1).getHost());
+        assertEquals(1000, serverAddresses.get(1).getPort());
+    }
+
+
+    @Test
+    public void shouldGetServerAddressForIPInMongoConnectionURLs() {
+        String mongoConnectionURLs = "172.28.32.156:1000";
+        List<ServerAddress> serverAddresses = MongoSinkFactoryUtil.getServerAddresses(mongoConnectionURLs, instrumentation);
+
+        assertEquals("172.28.32.156", serverAddresses.get(0).getHost());
+        assertEquals(1000, serverAddresses.get(0).getPort());
+    }
+
+    @Test
+    public void shouldThrowExceptionIfHostAndPortNotProvidedProperly() {
+        String mongoConnectionURLs = "test";
+        thrown.expect(IllegalArgumentException.class);
+        thrown.expectMessage("SINK_MONGO_CONNECTION_URLS should contain host and port both");
+
+        MongoSinkFactoryUtil.getServerAddresses(mongoConnectionURLs, instrumentation);
+    }
+
+    @Test
+    public void shouldReturnBlackListRetryStatusCodesAsList() {
+        String inputRetryStatusCodeBlacklist = "404, 502";
+        List<String> statusCodesAsList = MongoSinkFactoryUtil.getStatusCodesAsList(inputRetryStatusCodeBlacklist);
+        assertEquals("404", statusCodesAsList.get(0));
+        assertEquals("502", statusCodesAsList.get(1));
+    }
+
+    @Test
+    public void shouldReturnEmptyBlackListRetryStatusCodesAsEmptyList() {
+        String inputRetryStatusCodeBlacklist = "";
+        List<String> statusCodesAsList = MongoSinkFactoryUtil.getStatusCodesAsList(inputRetryStatusCodeBlacklist);
+        assertEquals(0, statusCodesAsList.size());
     }
 }
